Reworked the RK4 scheme...still not getting the right result in single precision. lets test it in double

+++ src/cl_kernels/useDouble.cl
@@ -4,7 +4,7 @@
 //typedef float FLOAT;
 //typedef float4 FLOAT4;
 
-typedef double FLOAT; 
+typedef float FLOAT; 
 typedef float4 FLOAT4;
 
 #endif
